pyproject.toml
[tool.poetry]
name = "hitfactorpy_graphql_server"
version = "0.0.3"
description = "Manage practical match reports with SQLAlchemy"
authors = ["Conor Heine <conor.heine@gmail.com>"]
readme = "README.md"
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Typing :: Typed",
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11"
]
include = [
  "hitfactorpy_graphql_server/py.typed",
]

[tool.poetry.scripts]
hitfactorpy-graphql-server = 'hitfactorpy_graphql_server.cli:cli'

[tool.poetry.dependencies]
python = "^3.10,<3.12"
typer = {version = "^0.7.0", extras = ["all"]}
hitfactorpy = "^0.1.1"
ariadne = "^0.17.1"
uvicorn = {version = "^0.20.0", extras = ["standard"]}
ariadne-graphql-modules = "^0.7.0"
hitfactorpy-sqlalchemy = "^0.0.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
flake8 = "^6.0.0"
black = "^22.12.0"
isort = "^5.11.4"
pre-commit = "^2.21.0"
tox = "^4.2.4"
mypy = "^0.991"
autoflake = "^2.0.0"
pytest-cov = "^4.0.0"
deptry = "^0.7.1"
coverage = "^7.0.5"
codecov = "^2.1.12"
mdx-include = "^1.4.2"
sqlalchemy-stubs = "^0.4"
pytest-docker = "^1.0.1"
docker-compose = "^1.29.2"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.10"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}
markdown-include = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
plugins = ["sqlalchemy.ext.mypy.plugin"]
mypy_path = "stubs"
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.black]
line-length = 120
target_version = ["py310"]
include = '\.pyi?$'
ignore_missing_imports = ["jinja2"]
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | build
  )/
)
'''

[tool.isort]
profile = "black"
src_paths = ["tests"]
known_first_party = ["hitfactorpy_graphql_server"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["hitfactorpy_graphql_server"]

[tool.deptry]
ignore_obsolete = [
  "asyncpg",
  "psycopg2-binary",
]
